requirements.txt
Flask==0.12.2
Flask-SocketIO==2.9.6
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==5.0.0
Werkzeug==0.12.2
click==6.7
itsdangerous==0.24
paho-mqtt==1.3.1
picamera==1.13
python-engineio==2.0.3
python-socketio==1.9.0
six==1.11.0
